Some small fixes to make work on newer 64-bit linux.  Warning: thebluealliance.com's API has changed, so some parts of the code wil have to be rewritten.

+++ set.h
@@ -77,7 +77,7 @@ std::set<T> to_set(std::vector<T> const& v){
 	return r;
 }
 
-template<typename T,unsigned LEN>
+template<typename T,size_t LEN>
 std::set<T> to_set(std::array<T,LEN> a){
 	std::set<T> r;
 	for(auto elem:a) r|=elem;
